added authorization for price rates

+++ app/controllers/company/hourly_rates_controller.rb
@@ -1,6 +1,8 @@
 class Company::HourlyRatesController < ApplicationController
 
   def index
+    authorize(current_company, :index_hourly_rates?)
+
     @base_rate = HourlyRate.where(company_id: current_company.id).base_rate
     @activity_price_rates = ImplicitHourlyRate
                                 .where(company_id: current_company.id)
@@ -11,6 +13,7 @@ class Company::HourlyRatesController < ApplicationController
 
   def create
     @hourly_rate = HourlyRate.new(hourly_rate_create_params.merge company: current_company)
+    authorize @hourly_rate
     if @hourly_rate.save
       flash[:success] = t 'flash.common.saved'
       redirect_to company_hourly_rates_path current_company
@@ -21,6 +24,7 @@ class Company::HourlyRatesController < ApplicationController
 
   def update
     @hourly_rate = HourlyRate.where(company: current_company).find(params[:id])
+    authorize @hourly_rate
     if @hourly_rate.update(hourly_rate_update_params)
       flash[:success] = t 'flash.common.saved'
       redirect_to company_hourly_rates_path current_company
@@ -31,6 +35,7 @@ class Company::HourlyRatesController < ApplicationController
 
   def destroy
     @hourly_rate = HourlyRate.find(params[:id])
+    authorize @hourly_rate
     if @hourly_rate.destroy
       flash[:success] = t 'flash.common.deleted'
       redirect_to company_hourly_rates_path current_company
